requirements.txt
Django>=5.0
psycopg2-binary>=2.9.9
pyHanko[image-support]>=0.21.0
cryptography>=42.0.0
python-dotenv>=1.0.1
gunicorn>=21.2.0
Pillow>=10.0.0
qrcode>=7.4.2
